Starting Agents Exercise
Exercise 1
Based on the lecture notes, Implement an Agent Class. The Agent should have a virtual update / draw function.

In order to allow movement, implement some basic physics variables to the agent, such as:

* Position

* Velocity

* Acceleration

* Force

And some basic Functions, Such as “AddForce”

In the agents update, calculate the appropriate new position of the agent.

With the above added, an agent should be movable simply by calling “AddForce” on an agent each frame.